third_party/blink/web_tests/external/wpt/close-watcher/basic.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="resources/helpers.js"></script>

<div id='d' style='height: 100px; width: 100px'></div>
<script>
test(t => {
  let events = [];
  let watcher = createRecordingCloseWatcher(t, events);

  watcher.requestClose();

  assert_array_equals(events, ["close"]);
}, "requestClose() with no user activation only fires close");

test(t => {
  let events = [];
  let watcher = createRecordingCloseWatcher(t, events);

  watcher.destroy();
  watcher.requestClose();

  assert_array_equals(events, []);
}, "destroy() then requestClose() fires no events");

test(t => {
  let events = [];
  let watcher = createRecordingCloseWatcher(t, events);

  watcher.close();
  assert_array_equals(events, ["close"]);

  watcher.requestClose();
  assert_array_equals(events, ["close"]);
}, "close() then requestClose() fires only one close event");

test(t => {
  let events = [];
  let watcher = createRecordingCloseWatcher(t, events);

  watcher.requestClose();
  assert_array_equals(events, ["close"]);

  watcher.destroy();
  assert_array_equals(events, ["close"]);
}, "requestClose() then destroy() fires only one close event");

test(t => {
  let events = [];
  let watcher = createRecordingCloseWatcher(t, events);

  watcher.close();
  assert_array_equals(events, ["close"]);

  watcher.destroy();
  assert_array_equals(events, ["close"]);
}, "close() then destroy() fires only one close event");

promise_test(async t => {
  let events = [];
  let watcher = createRecordingCloseWatcher(t, events);

  await sendCloseSignal();

  assert_array_equals(events, ["close"]);
}, "Esc key does not count as user activation, so it fires close but not cancel");

promise_test(async t => {
  let events = [];
  let watcher = createRecordingCloseWatcher(t, events);

  watcher.destroy();
  await sendCloseSignal();

  assert_array_equals(events, []);
}, "destroy() then close via Esc key fires no events");

promise_test(async t => {
  let events = [];
  let watcher = createRecordingCloseWatcher(t, events);

  await sendCloseSignal();
  watcher.destroy();

  assert_array_equals(events, ["close"]);
}, "Esc key then destroy() fires only one close event");

test(t => {
  let events = [];
  let watcher = createRecordingCloseWatcher(t, events);

  let keydown = new KeyboardEvent('keydown', {'key': 'Escape', 'keyCode': 27});
  window.dispatchEvent(keydown);
  let keyup = new KeyboardEvent('keyup', {'key': 'Escape', 'keyCode': 27});
  window.dispatchEvent(keyup);

  assert_array_equals(events, []);

  let keyup2 = document.createEvent("Event");
  keyup2.initEvent("keyup", true);
  window.dispatchEvent(keyup2);

  assert_array_equals(events, []);
}, "close via synthesized Esc key should not work");
</script>
